move JDBC/H2 backend implementation to separate module
