+++ oshdb-api-jdbc/src/main/java/org/heigit/ohsome/oshdb/api/db/OSHDBH2.java
