+++ oshdb-api-jdbc/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestFlatMapAggregateGroupedByEntity.java
@@ -4,26 +4,24 @@ import static org.junit.Assert.assertEquals;
 
 import java.util.AbstractMap.SimpleImmutableEntry;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map.Entry;
-import java.util.Set;
 import java.util.SortedMap;
 import org.heigit.ohsome.oshdb.OSHDBBoundingBox;
-import org.heigit.ohsome.oshdb.OSHDBTag;
 import org.heigit.ohsome.oshdb.api.db.OSHDBDatabase;
 import org.heigit.ohsome.oshdb.api.db.OSHDBH2;
 import org.heigit.ohsome.oshdb.api.mapreducer.MapReducer;
 import org.heigit.ohsome.oshdb.api.mapreducer.OSMContributionView;
 import org.heigit.ohsome.oshdb.osm.OSMType;
+import org.heigit.ohsome.oshdb.util.celliterator.ContributionType;
 import org.heigit.ohsome.oshdb.util.mappable.OSMContribution;
 import org.heigit.ohsome.oshdb.util.time.OSHDBTimestamps;
 import org.junit.Test;
 
 /**
- * Test flat map method of the MapAggregator class of the OSHDB API.
+ * Test flat map method with groupByEntity of the MapAggregator class of the OSHDB API.
  */
-public class TestFlatMapAggregate {
+public class TestFlatMapAggregateGroupedByEntity {
   private final OSHDBDatabase oshdb;
 
   private final OSHDBBoundingBox bbox = new OSHDBBoundingBox(8, 49, 9, 50);
@@ -32,8 +30,8 @@ public class TestFlatMapAggregate {
 
   private static final double DELTA = 1e-8;
 
-  public TestFlatMapAggregate() throws Exception {
-    oshdb = new OSHDBH2("./src/test/resources/test-data");
+  public TestFlatMapAggregateGroupedByEntity() throws Exception {
+    oshdb = new OSHDBH2("../oshdb-api/src/test/resources/test-data");
   }
 
   private MapReducer<OSMContribution> createMapReducerOSMContribution() throws Exception {
@@ -46,39 +44,36 @@ public class TestFlatMapAggregate {
 
   @Test
   public void test() throws Exception {
-    SortedMap<Long, Set<Entry<Integer, Integer>>> result = createMapReducerOSMContribution()
+    SortedMap<Long, Integer> result = createMapReducerOSMContribution()
         .timestamps(timestamps72)
-        .flatMap(
-            contribution -> {
-              if (contribution.getEntityAfter().getId() != 617308093) {
-                  return new ArrayList<>();
-                }
-              List<Entry<Long, Entry<Integer, Integer>>> ret = new ArrayList<>();
-              for (OSHDBTag tag : contribution.getEntityAfter().getTags()) {
-                ret.add(new SimpleImmutableEntry<>(
-                    contribution.getEntityAfter().getId(),
-                    new SimpleImmutableEntry<>(tag.getKey(), tag.getValue())
-                ));
-              }
-              return ret;
-            }
-        )
+        .groupByEntity()
+        .flatMap(contributions -> {
+          if (contributions.get(0).getEntityAfter().getId() != 617308093) {
+            return new ArrayList<>();
+          }
+          List<Entry<Long, Integer>> ret = new ArrayList<>();
+          ret.add(new SimpleImmutableEntry<>(
+              contributions.get(0).getEntityAfter().getId(),
+              (int) contributions
+                  .stream()
+                  .filter(c -> c.getContributionTypes().contains(ContributionType.GEOMETRY_CHANGE))
+                  .count()
+          ));
+          ret.add(new SimpleImmutableEntry<>(
+              contributions.get(0).getEntityAfter().getId(),
+              2
+          ));
+          return ret;
+        })
         .aggregateBy(Entry::getKey)
         .map(Entry::getValue)
         .reduce(
-            HashSet::new,
-            (x, y) -> {
-              x.add(y);
-              return x;
-            },
-            (x, y) -> {
-              Set<Entry<Integer, Integer>> ret = new HashSet<>(x);
-              ret.addAll(y);
-              return ret;
-            }
+            () -> 0,
+            Integer::sum,
+            Integer::sum
         );
 
     assertEquals(1, result.entrySet().size());
-    assertEquals(2, result.get(617308093L).size());
+    assertEquals(5 + 2, result.get(617308093L).intValue());
   }
 }
