+++ oshdb-api-jdbc/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestFlatMapReduceGroupedByEntityOSHDBH2Multithread.java
@@ -7,15 +7,16 @@ import org.heigit.ohsome.oshdb.api.db.OSHDBH2;
  *
  * <p>Runs the tests using the multithreaded H2 backend.</p>
  */
-public class TestMapReduceOSHDBH2Multithread extends TestMapReduce {
+public class TestFlatMapReduceGroupedByEntityOSHDBH2Multithread extends
+    TestFlatMapReduceGroupedByEntity {
   /**
    * Creates the test runner using the multithreaded H2 backend.
    *
    * @throws Exception if something goes wrong
    */
-  public TestMapReduceOSHDBH2Multithread() throws Exception {
+  public TestFlatMapReduceGroupedByEntityOSHDBH2Multithread() throws Exception {
     super(
-        (new OSHDBH2("./src/test/resources/test-data")).multithreading(true)
+        (new OSHDBH2("../oshdb-api/src/test/resources/test-data")).multithreading(true)
     );
   }
 }
