+++ oshdb-api-jdbc/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestLambdaFilter.java
@@ -28,7 +28,7 @@ public class TestLambdaFilter {
   private static final double DELTA = 1e-8;
 
   public TestLambdaFilter() throws Exception {
-    oshdb = new OSHDBH2("./src/test/resources/test-data");
+    oshdb = new OSHDBH2("../oshdb-api/src/test/resources/test-data");
   }
 
   private MapReducer<OSMContribution> createMapReducerOSMContribution() throws Exception {
