+++ oshdb-api-jdbc/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestMapAggregateByGeometry.java
@@ -39,7 +39,7 @@ public class TestMapAggregateByGeometry {
   private static final double DELTA = 1e-4;
 
   public TestMapAggregateByGeometry() throws Exception {
-    oshdb = new OSHDBH2("./src/test/resources/test-data");
+    oshdb = new OSHDBH2("../oshdb-api/src/test/resources/test-data");
   }
 
   private MapReducer<OSMContribution> createMapReducerOSMContribution() throws Exception {
