+++ oshdb-api-jdbc/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestMapAggregateByTimestamp.java
@@ -36,7 +36,7 @@ public class TestMapAggregateByTimestamp {
   private static final double DELTA = 1e-8;
 
   public TestMapAggregateByTimestamp() throws Exception {
-    oshdb = new OSHDBH2("./src/test/resources/test-data");
+    oshdb = new OSHDBH2("../oshdb-api/src/test/resources/test-data");
   }
 
   private MapReducer<OSMContribution> createMapReducerOSMContribution() throws Exception {
