+++ oshdb-api/src/main/java/org/heigit/ohsome/oshdb/api/db/OSHDBKeytablesConnection.java
@@ -1,0 +1,10 @@
+package org.heigit.ohsome.oshdb.api.db;
+
+import java.sql.Connection;
+
+/**
+ * An interface for providing access to a keytables database.
+ */
+public interface OSHDBKeytablesConnection {
+  Connection getConnection();
+}
