+++ pom.xml
@@ -27,6 +27,7 @@
     <module>oshdb-filter</module>
     <module>oshdb-api</module>
     <module>oshdb-api-ignite</module>
+    <module>oshdb-api-jdbc</module>
   </modules>
 
   <properties>
